+++ src/visits/ShortUrlVisits.js
@@ -1,8 +1,9 @@
-import { isEmpty, values } from 'ramda';
-import React, { useState, useEffect } from 'react';
+import { isEmpty, propEq, values } from 'ramda';
+import React, { useState, useEffect, useMemo } from 'react';
 import { Button, Card, Collapse } from 'reactstrap';
 import PropTypes from 'prop-types';
 import qs from 'qs';
+import classNames from 'classnames';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faChevronDown as chevronDown } from '@fortawesome/free-solid-svg-icons';
 import DateRangeRow from '../utils/DateRangeRow';
@@ -41,10 +42,9 @@ const highlightedVisitsToStats = (highlightedVisits, prop) => highlightedVisits.
   return acc;
 }, {});
 const format = formatDate();
-let memoizationId;
-let timeWhenMounted;
+let selectedBar;
 
-const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
+const ShortUrlVisits = ({ processStatsFromVisits, normalizeVisits }, OpenMapModalBtn) => {
   const ShortUrlVisitsComp = ({
     match,
     location,
@@ -62,23 +62,40 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
     const [ highlightedVisits, setHighlightedVisits ] = useState([]);
     const [ isMobileDevice, setIsMobileDevice ] = useState(false);
     const determineIsMobileDevice = () => setIsMobileDevice(matchMedia('(max-width: 991px)').matches);
+    const setSelectedVisits = (selectedVisits) => {
+      selectedBar = undefined;
+      setHighlightedVisits(selectedVisits);
+    };
+    const highlightVisitsForProp = (prop) => (value) => {
+      const newSelectedBar = `${prop}_${value}`;
+
+      if (selectedBar === newSelectedBar) {
+        setHighlightedVisits([]);
+        selectedBar = undefined;
+      } else {
+        setHighlightedVisits(normalizedVisits.filter(propEq(prop, value)));
+        selectedBar = newSelectedBar;
+      }
+    };
 
     const { params } = match;
     const { shortCode } = params;
     const { search } = location;
     const { domain } = qs.parse(search, { ignoreQueryPrefix: true });
 
-    const loadVisits = () => {
-      const start = format(startDate);
-      const end = format(endDate);
+    const { visits, loading, loadingLarge, error } = shortUrlVisits;
+    const showTableControls = !loading && visits.length > 0;
+    const normalizedVisits = useMemo(() => normalizeVisits(visits), [ visits ]);
+    const { os, browsers, referrers, countries, cities, citiesForMap } = useMemo(
+      () => processStatsFromVisits(normalizedVisits),
+      [ normalizedVisits ]
+    );
+    const mapLocations = values(citiesForMap);
 
-      // While the "page" is loaded, use the timestamp + filtering dates as memoization IDs for stats calculations
-      memoizationId = `${timeWhenMounted}_${shortCode}_${start}_${end}`;
-      getShortUrlVisits(shortCode, { startDate: start, endDate: end, domain });
-    };
+    const loadVisits = () =>
+      getShortUrlVisits(shortCode, { startDate: format(startDate), endDate: format(endDate), domain });
 
     useEffect(() => {
-      timeWhenMounted = new Date().getTime();
       getShortUrlDetail(shortCode, domain);
       determineIsMobileDevice();
       window.addEventListener('resize', determineIsMobileDevice);
@@ -91,9 +108,6 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
     useEffect(() => {
       loadVisits();
     }, [ startDate, endDate ]);
-
-    const { visits, loading, loadingLarge, error } = shortUrlVisits;
-    const showTableControls = !loading && visits.length > 0;
 
     const renderVisitsContent = () => {
       if (loading) {
@@ -114,11 +128,6 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
         return <Message>There are no visits matching current filter  :(</Message>;
       }
 
-      const { os, browsers, referrers, countries, cities, citiesForMap } = processStatsFromVisits(
-        { id: memoizationId, visits }
-      );
-      const mapLocations = values(citiesForMap);
-
       return (
         <div className="row">
           <div className="col-xl-4 col-lg-6">
@@ -137,6 +146,7 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
                 name: 'Referrer name',
                 amount: 'Visits amount',
               }}
+              onClick={highlightVisitsForProp('referer')}
             />
           </div>
           <div className="col-lg-6">
@@ -148,6 +158,7 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
                 name: 'Country name',
                 amount: 'Visits amount',
               }}
+              onClick={highlightVisitsForProp('country')}
             />
           </div>
           <div className="col-lg-6">
@@ -163,6 +174,7 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
                 name: 'City name',
                 amount: 'Visits amount',
               }}
+              onClick={highlightVisitsForProp('city')}
             />
           </div>
         </div>
@@ -175,24 +187,35 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
 
         <section className="mt-4">
           <div className="row flex-md-row-reverse">
-            <div className="col-lg-8 col-xl-6">
+            <div className="col-lg-7 col-xl-6">
               <DateRangeRow
+                disabled={loading}
                 startDate={startDate}
                 endDate={endDate}
                 onStartDateChange={setStartDate}
                 onEndDateChange={setEndDate}
               />
             </div>
-            <div className="col-lg-4 col-xl-6 mt-4 mt-lg-0">
+            <div className="col-lg-5 col-xl-6 mt-4 mt-lg-0">
               {showTableControls && (
-                <Button
-                  outline
-                  block={isMobileDevice}
-                  onClick={toggleTable}
-                >
-                  {showTable ? 'Hide' : 'Show'} table{' '}
-                  <FontAwesomeIcon icon={chevronDown} rotation={showTable ? 180 : undefined} />
-                </Button>
+                <span className={classNames({ 'row flex-row-reverse': isMobileDevice })}>
+                  <span className={classNames({ 'col-6': isMobileDevice })}>
+                    <Button outline color="primary" block={isMobileDevice} onClick={toggleTable}>
+                      {showTable ? 'Hide' : 'Show'} table
+                      <FontAwesomeIcon icon={chevronDown} rotation={showTable ? 180 : undefined} className="ml-2" />
+                    </Button>
+                  </span>
+                  <span className={classNames({ 'col-6': isMobileDevice, 'ml-2': !isMobileDevice })}>
+                    <Button
+                      outline
+                      disabled={highlightedVisits.length === 0}
+                      block={isMobileDevice}
+                      onClick={() => setSelectedVisits([])}
+                    >
+                      Reset selection
+                    </Button>
+                  </span>
+                </span>
               )}
             </div>
           </div>
@@ -201,12 +224,16 @@ const ShortUrlVisits = ({ processStatsFromVisits }, OpenMapModalBtn) => {
         {showTableControls && (
           <Collapse
             isOpen={showTable}
-
             // Enable stickiness only when there's no CSS animation, to avoid weird rendering effects
             onEntered={setSticky}
             onExiting={unsetSticky}
           >
-            <VisitsTable visits={visits} isSticky={tableIsSticky} onVisitsSelected={setHighlightedVisits} />
+            <VisitsTable
+              visits={normalizedVisits}
+              selectedVisits={highlightedVisits}
+              setSelectedVisits={setSelectedVisits}
+              isSticky={tableIsSticky}
+            />
           </Collapse>
         )}